+++ Models/ViewModels/ChildCareViewModel.cs
@@ -7,7 +7,7 @@ namespace Models.ViewModels
     {
         public long Id { get; set; }
         public long UserId { get; set; }
-        public int ChildId { get; set; }
+        public long ChildId { get; set; }
         public int SchoolFather { get; set; }
         public int SchoolMother { get; set; }
         public int SchoolNonParent { get; set; }
